Client login & onboarding

+++ fastfx-recharge-apis/src/main/java/com/kavki/fastfxrechargeapis/DAO/AdminRepositories/PrefundRepo.java
@@ -4,7 +4,7 @@ import com.kavki.fastfxrechargeapis.Entity.Admin.PrefundEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface PrefundRepo extends JpaRepository<PrefundEntity, Long>{
+public interface PrefundRepo extends JpaRepository<PrefundEntity, Integer>{
 
     @Query(value="select sum(c.Account_balance) from client_records c",nativeQuery = true)
     Double calcTotal();
